Converts express-graphql into apollo-server-express
Change the api server to use apollo-server instead of express-graphql
