+++ package.json
@@ -39,6 +39,7 @@
     "webpack": "^3.8.1"
   },
   "dependencies": {
+    "apollo-server-express": "^1.3.0",
     "body-parser": "^1.18.2",
     "chalk": "^2.3.0",
     "cheerio": "^1.0.0-rc.2",
@@ -46,8 +47,8 @@
     "cookie-parser": "^1.4.3",
     "dotenv": "^4.0.0",
     "express": "^4.16.2",
-    "express-graphql": "^0.6.11",
     "graphql": "^0.11.7",
+    "graphql-tools": "^2.13.0",
     "moment": "^2.19.4",
     "node-persist": "^2.1.0",
     "node-storage": "0.0.7",
